+++ CHANGELOG.md
@@ -8,6 +8,9 @@ All notable changes to this project.
 - The minimum required version of `aiohttp` is now `3.8.1`.
 - The minimum required version of `ujson` is now `5.2.0`.
 
+### Removed
+- Tornado support: `TornadoExecutor` and `TornadoConnector` classes.
+
 
 ## [3.7.0] - 2022-04-22